+++ packages/react-button/src/components/Button/Button.types.ts
@@ -1,9 +1,7 @@
 import * as React from 'react';
 import { BaseSlots, SlotProps } from '@fluentui/react-compose';
 import { ComponentProps, ShorthandProps } from '@fluentui/react-compose/lib/next/index';
-import { ColorTokenSet, RecursivePartial } from '@fluentui/theme';
-
-export type SizeValue = 'smallest' | 'smaller' | 'small' | 'medium' | 'large' | 'larger' | 'largest';
+import { ColorTokenSet, RecursivePartial, SizeValue } from '@fluentui/theme';
 
 export type ButtonProps = ComponentProps &
   React.HTMLAttributes<HTMLElement> & {