+++ packages/react-next/src/components/Coachmark/Coachmark.doc.tsx
@@ -3,12 +3,11 @@ import * as React from 'react';
 import { CoachmarkBasicExample } from './examples/Coachmark.Basic.Example';
 
 import { IDocPageProps } from '../../common/DocPage.types';
-const CoachmarkBasicExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Coachmark/examples/Coachmark.Basic.Example.tsx') as string;
+const CoachmarkBasicExampleCode = require('!raw-loader!@fluentui/react-next/src/components/Coachmark/examples/Coachmark.Basic.Example.tsx') as string;
 export const CoachmarkPageProps: IDocPageProps = {
   title: 'Coachmark',
   componentName: 'Coachmark',
-  componentUrl:
-    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/Coachmark',
+  componentUrl: 'https://github.com/microsoft/fluentui/tree/master/packages/react-next/src/components/Coachmark',
   examples: [
     {
       title: 'Coachmark Basic',
@@ -17,10 +16,10 @@ export const CoachmarkPageProps: IDocPageProps = {
       isScrollable: false,
     },
   ],
-  overview: require<string>('!raw-loader!office-ui-fabric-react/src/components/Coachmark/docs/CoachmarkOverview.md'),
+  overview: require<string>('!raw-loader!@fluentui/react-next/src/components/Coachmark/docs/CoachmarkOverview.md'),
   bestPractices: '',
-  dos: require<string>('!raw-loader!office-ui-fabric-react/src/components/Coachmark/docs/CoachmarkDos.md'),
-  donts: require<string>('!raw-loader!office-ui-fabric-react/src/components/Coachmark/docs/CoachmarkDonts.md'),
+  dos: require<string>('!raw-loader!@fluentui/react-next/src/components/Coachmark/docs/CoachmarkDos.md'),
+  donts: require<string>('!raw-loader!@fluentui/react-next/src/components/Coachmark/docs/CoachmarkDonts.md'),
   isHeaderVisible: true,
   isFeedbackVisible: true,
 };