+++ packages/react-next/src/components/MessageBar/MessageBar.doc.tsx
@@ -4,13 +4,12 @@ import { IDocPageProps } from '../../common/DocPage.types';
 
 import { MessageBarBasicExample } from './examples/MessageBar.Basic.Example';
 
-const MessageBarBasicExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/MessageBar/examples/MessageBar.Basic.Example.tsx') as string;
+const MessageBarBasicExampleCode = require('!raw-loader!@fluentui/react-next/src/components/MessageBar/examples/MessageBar.Basic.Example.tsx') as string;
 
 export const MessageBarPageProps: IDocPageProps = {
   title: 'MessageBar',
   componentName: 'MessageBar',
-  componentUrl:
-    'https://github.com/microsoft/fluentui/tree/master/packages/office-ui-fabric-react/src/components/MessageBar',
+  componentUrl: 'https://github.com/microsoft/fluentui/tree/master/packages/react-next/src/components/MessageBar',
   examples: [
     {
       title: 'Various MessageBar types',
@@ -18,10 +17,10 @@ export const MessageBarPageProps: IDocPageProps = {
       view: <MessageBarBasicExample />,
     },
   ],
-  overview: require<string>('!raw-loader!office-ui-fabric-react/src/components/MessageBar/docs/MessageBarOverview.md'),
+  overview: require<string>('!raw-loader!@fluentui/react-next/src/components/MessageBar/docs/MessageBarOverview.md'),
   bestPractices: require<
     string
-  >('!raw-loader!office-ui-fabric-react/src/components/MessageBar/docs/MessageBarBestPractices.md'),
+  >('!raw-loader!@fluentui/react-next/src/components/MessageBar/docs/MessageBarBestPractices.md'),
   isHeaderVisible: true,
   isFeedbackVisible: true,
 };