+++ packages/react-next/CHANGELOG.md
@@ -1,8 +1,29 @@
 # Change Log - @fluentui/react-next
 
-This log was last generated on Thu, 17 Sep 2020 23:15:07 GMT and should not be manually modified.
+This log was last generated on Tue, 22 Sep 2020 12:25:45 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.0.0-alpha.114](https://github.com/microsoft/fluentui/tree/@fluentui/react-next_v8.0.0-alpha.114)
+
+Tue, 22 Sep 2020 12:25:45 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-next_v8.0.0-alpha.113..@fluentui/react-next_v8.0.0-alpha.114)
+
+### Changes
+
+- Updating Coachmark.Basic.Example.tsx to a function component. ([PR #15104](https://github.com/microsoft/fluentui/pull/15104) by czearing@outlook.com)
+- Updating react-next component documentation to the newest versions found in office-ui-fabric-react. ([PR #15107](https://github.com/microsoft/fluentui/pull/15107) by czearing@outlook.com)
+- Autofill: Add to react-next ([PR #15148](https://github.com/microsoft/fluentui/pull/15148) by joschect@microsoft.com)
+- Updating react-next example imports. ([PR #15153](https://github.com/microsoft/fluentui/pull/15153) by czearing@outlook.com)
+
+## [8.0.0-alpha.113](https://github.com/microsoft/fluentui/tree/@fluentui/react-next_v8.0.0-alpha.113)
+
+Mon, 21 Sep 2020 12:22:29 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-next_v8.0.0-alpha.112..@fluentui/react-next_v8.0.0-alpha.113)
+
+### Changes
+
+- Updating Layer to a function component. ([PR #14698](https://github.com/microsoft/fluentui/pull/14698) by czearing@outlook.com)
 
 ## [8.0.0-alpha.111](https://github.com/microsoft/fluentui/tree/@fluentui/react-next_v8.0.0-alpha.111)